routes working and modulating code WIP
