+++ src/utils/animationLogo.tsx
@@ -24,7 +24,7 @@ export const animationLogo = ({
     copyButtonRef,
     tl
 }: AnimationLogoProps) => {
-
+  console.log("copyButtonRef: ", copyButtonRef);
     if(startRef.current
         && startTextRef.current
         && firstPipeRef.current
